backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pandas
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pytest
